Fixed getState and setEpisodeStartTime in Java binding, buffers type changed to byte[]
Former-commit-id: 504afe9da78e0a17c7604b97a717f5ae247ed4a2

+++ src/lib_java/java_classes/GameState.java
@@ -8,19 +8,19 @@ public class GameState{
     public int number;
     public int[] gameVariables;
 
-    public int[] screenBuffer;
-    public int[] depthBuffer;
-    public int[] labelsBuffer;
-    public int[] automapBuffer;
+    public byte[] screenBuffer;
+    public byte[] depthBuffer;
+    public byte[] labelsBuffer;
+    public byte[] automapBuffer;
 
     public Label[] labels;
 
     GameState(int number,
         int[] gameVariables,
-        int[] screenBuffer,
-        int[] depthBuffer,
-        int[] labelsBuffer,
-        int[] automapBuffer,
+        byte[] screenBuffer,
+        byte[] depthBuffer,
+        byte[] labelsBuffer,
+        byte[] automapBuffer,
         Label[] labels){
 
         this.number = number;
